Refactor ModelStub in tests
Pulled the ModelStub class out from existing tests for reuse
across other tests and to improve maintainability.
